\name{buildSummaryTable}
\alias{buildSummaryTable}

\title{
	Build Summary Table For All Examples From Literature
}
\description{
	Build Summary Table For All Examples From Literature
}
\usage{
	buildSummaryTable()
}
\details{
	See also the documentation for the data files.
}
\value{
  TODO
}
\author{
  Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\references{
  See the documentation for the data files.
}
\keyword{ misc }

\examples{
  buildSummaryTable()
}