% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order.R
\name{canonicalOrder}
\alias{canonicalOrder}
\title{Sorts sequences of a design into a canonical order}
\usage{
canonicalOrder(design)
}
\arguments{
\item{design}{Cross-over design.}
}
\description{
Sorts sequences of a design into a canonical order.
}
\details{
When comparing bigger designs this ordering easily allows to check whether two
designs are equal.
}
\examples{
getDesign("switchback5t")
canonicalOrder(getDesign("switchback5t"))
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{misc}

