% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CrossoverDesign-class}
\alias{CrossoverDesign}
\alias{CrossoverDesign-class}
\alias{print,CrossoverDesign-method}
\alias{show,CrossoverDesign-method}
\title{Class CrossoverDesign}
\description{
A S4 class for Crossover designs: CrossoverDesign
}
\section{Slots}{
 \describe{ \item{list("design")}{ Matrix }\item{:}{ Matrix }
\item{list("s")}{ Number of sequences. }\item{:}{ Number of sequences. }
\item{list("p")}{ Number of periods. }\item{:}{ Number of periods. }
\item{list("v")}{ Number of treatments. }\item{:}{ Number of treatments. }
\item{list("model")}{ A numeric specifying the model the design was searched
for or -1 if unknown. }\item{:}{ A numeric specifying the model the design
was searched for or -1 if unknown. } \item{list("description")}{ Optional
description of design or reference. }\item{:}{ Optional description of
design or reference. } \item{list("attr")}{ List with attributes. }\item{:}{
List with attributes. } \item{list("misc")}{ List with miscellaneous stuff -
not used yet. }\item{:}{ List with miscellaneous stuff - not used yet. } }
}
\examples{
design <- t(rbind(c(1,1,2,2),
                  c(2,2,1,1),
                  c(1,1,2,2),
                  c(2,2,1,1),
                  c(1,2,2,1),
                  c(2,1,1,2)))

new("CrossoverDesign", design)
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{graphs}

