\docType{data}
\name{Crossoverdesigns}
\alias{Crossoverdesigns}
\alias{anderson}
\alias{andersonPreece}
\alias{archdeacon}
\alias{atkinson3t}
\alias{atkinson4t}
\alias{atkinson5t}
\alias{balaam3t}
\alias{balaam4t}
\alias{balaam5t}
\alias{balaam6t}
\alias{bateJones5t}
\alias{bateJones8t}
\alias{berenblut3t}
\alias{berenblut4t}
\alias{berenblut5t}
\alias{blaisdellRaghavarao6t}
\alias{blaisdellRaghavarao8t}
\alias{blaisdellRaghavarao9t}
\alias{davisHall6ta}
\alias{davisHall6tb}
\alias{davisHall6tc}
\alias{davisHall7ta}
\alias{davisHall7tb}
\alias{davisHall7tc}
\alias{davisHall8ta}
\alias{davisHall8tb}
\alias{davisHall8tc}
\alias{davisHall9ta}
\alias{davisHall9tb}
\alias{davisHall9tc}
\alias{federerAtkinson3ta}
\alias{federerAtkinson3tb}
\alias{federerAtkinson4ta}
\alias{federerAtkinson4tb}
\alias{federerAtkinson5ta}
\alias{fletcher1}
\alias{fletcher10}
\alias{fletcher11}
\alias{fletcher12}
\alias{fletcher13}
\alias{fletcher14}
\alias{fletcher15}
\alias{fletcher16}
\alias{fletcher17}
\alias{fletcher18}
\alias{fletcher19}
\alias{fletcher2}
\alias{fletcher20}
\alias{fletcher21}
\alias{fletcher22}
\alias{fletcher23}
\alias{fletcher24}
\alias{fletcher25}
\alias{fletcher26}
\alias{fletcher27}
\alias{fletcher28}
\alias{fletcher29}
\alias{fletcher3}
\alias{fletcher30}
\alias{fletcher31}
\alias{fletcher4}
\alias{fletcher5}
\alias{fletcher6}
\alias{fletcher7}
\alias{fletcher8}
\alias{fletcher9}
\alias{iqbalJones1}
\alias{iqbalJones10}
\alias{iqbalJones11}
\alias{iqbalJones12}
\alias{iqbalJones13}
\alias{iqbalJones14}
\alias{iqbalJones15}
\alias{iqbalJones16}
\alias{iqbalJones17}
\alias{iqbalJones18}
\alias{iqbalJones19}
\alias{iqbalJones2}
\alias{iqbalJones20}
\alias{iqbalJones21}
\alias{iqbalJones22}
\alias{iqbalJones23}
\alias{iqbalJones24}
\alias{iqbalJones25}
\alias{iqbalJones26}
\alias{iqbalJones27}
\alias{iqbalJones28}
\alias{iqbalJones29}
\alias{iqbalJones3}
\alias{iqbalJones30}
\alias{iqbalJones31}
\alias{iqbalJones32}
\alias{iqbalJones33}
\alias{iqbalJones34}
\alias{iqbalJones35}
\alias{iqbalJones36}
\alias{iqbalJones37}
\alias{iqbalJones38}
\alias{iqbalJones39}
\alias{iqbalJones4}
\alias{iqbalJones40}
\alias{iqbalJones41}
\alias{iqbalJones42}
\alias{iqbalJones5}
\alias{iqbalJones6}
\alias{iqbalJones7}
\alias{iqbalJones8}
\alias{iqbalJones9}
\alias{lewisFletcherMatthews1}
\alias{lewisFletcherMatthews10}
\alias{lewisFletcherMatthews11}
\alias{lewisFletcherMatthews12}
\alias{lewisFletcherMatthews13}
\alias{lewisFletcherMatthews14}
\alias{lewisFletcherMatthews15}
\alias{lewisFletcherMatthews16}
\alias{lewisFletcherMatthews17}
\alias{lewisFletcherMatthews18}
\alias{lewisFletcherMatthews19}
\alias{lewisFletcherMatthews2}
\alias{lewisFletcherMatthews20}
\alias{lewisFletcherMatthews3}
\alias{lewisFletcherMatthews4}
\alias{lewisFletcherMatthews5}
\alias{lewisFletcherMatthews6}
\alias{lewisFletcherMatthews7}
\alias{lewisFletcherMatthews8}
\alias{lewisFletcherMatthews9}
\alias{orthogonalLatinSquare3t}
\alias{orthogonalLatinSquare4t}
\alias{orthogonalLatinSquare5t}
\alias{orthogonalLatinSquare7t}
\alias{pattersonLucasExtraPeriod30}
\alias{pattersonLucasExtraPeriod31}
\alias{pattersonLucasExtraPeriod32}
\alias{pattersonLucasExtraPeriod33}
\alias{pattersonLucasExtraPeriod34}
\alias{pattersonLucasExtraPeriod35}
\alias{pattersonLucasExtraPeriod36}
\alias{pattersonLucasExtraPeriod37}
\alias{pattersonLucasExtraPeriod38}
\alias{pattersonLucasExtraPeriod39}
\alias{pattersonLucasExtraPeriod40}
\alias{pattersonLucasExtraPeriod41}
\alias{pattersonLucasExtraPeriod42}
\alias{pattersonLucasExtraPeriod43}
\alias{pattersonLucasExtraPeriod44}
\alias{pattersonLucasExtraPeriod45}
\alias{pattersonLucasExtraPeriod46}
\alias{pattersonLucasExtraPeriod47}
\alias{pattersonLucasExtraPeriod48}
\alias{pattersonLucasExtraPeriod49}
\alias{pattersonLucasExtraPeriod86}
\alias{pattersonLucasPBIBD100}
\alias{pattersonLucasPBIBD101}
\alias{pattersonLucasPBIBD102}
\alias{pattersonLucasPBIBD103}
\alias{pattersonLucasPBIBD104}
\alias{pattersonLucasPBIBD105}
\alias{pattersonLucasPBIBD106}
\alias{pattersonLucasPBIBD107}
\alias{pattersonLucasPBIBD125}
\alias{pattersonLucasPBIBD126}
\alias{pattersonLucasPBIBD127}
\alias{pattersonLucasPBIBD128}
\alias{pattersonLucasPBIBD131}
\alias{pattersonLucasPBIBD132}
\alias{pattersonLucasPBIBD133}
\alias{pattersonLucasPBIBD134}
\alias{pattersonLucasPBIBD135}
\alias{pattersonLucasPBIBD136}
\alias{pattersonLucasPBIBD137}
\alias{pattersonLucasPBIBD138}
\alias{pattersonLucasPBIBD139}
\alias{pattersonLucasPBIBD140}
\alias{pattersonLucasPBIBD141}
\alias{pattersonLucasPBIBD153}
\alias{pattersonLucasPBIBD154}
\alias{pattersonLucasPBIBD155}
\alias{pattersonLucasPBIBD156}
\alias{pattersonLucasPBIBD99}
\alias{pattersonLucasPltT1}
\alias{pattersonLucasPltT10}
\alias{pattersonLucasPltT12}
\alias{pattersonLucasPltT13}
\alias{pattersonLucasPltT15}
\alias{pattersonLucasPltT16}
\alias{pattersonLucasPltT17}
\alias{pattersonLucasPltT18}
\alias{pattersonLucasPltT19}
\alias{pattersonLucasPltT20}
\alias{pattersonLucasPltT21}
\alias{pattersonLucasPltT22}
\alias{pattersonLucasPltT23}
\alias{pattersonLucasPltT3}
\alias{pattersonLucasPltT4}
\alias{pattersonLucasPltT5}
\alias{pattersonLucasPltT7}
\alias{pattersonLucasPltT8}
\alias{pattersonLucasPltT9}
\alias{pidgeon1}
\alias{pidgeon10}
\alias{pidgeon11}
\alias{pidgeon12}
\alias{pidgeon13}
\alias{pidgeon14}
\alias{pidgeon15}
\alias{pidgeon16}
\alias{pidgeon17}
\alias{pidgeon18}
\alias{pidgeon19}
\alias{pidgeon2}
\alias{pidgeon20}
\alias{pidgeon3}
\alias{pidgeon4}
\alias{pidgeon5}
\alias{pidgeon6}
\alias{pidgeon7}
\alias{pidgeon8}
\alias{pidgeon9}
\alias{prescott1}
\alias{prescott2}
\alias{quenouille3t1}
\alias{quenouille3t2}
\alias{quenouille4t1}
\alias{quenouille4t2}
\alias{quenouille4t3}
\alias{russel4t}
\alias{russel7t}
\alias{switchback3t}
\alias{switchback4t}
\alias{switchback5t}
\alias{switchback6t}
\alias{switchback7t}
\alias{williams3t}
\alias{williams4t}
\alias{williams5t}
\alias{williams6t}
\alias{williams7t}
\alias{williams8t}
\alias{williams9t}
\alias{clatworthy.r1}
\alias{clatworthy.r10}
\alias{clatworthy.r101}
\alias{clatworthy.r103}
\alias{clatworthy.r104}
\alias{clatworthy.r106}
\alias{clatworthy.r109}
\alias{clatworthy.r111}
\alias{clatworthy.r112}
\alias{clatworthy.r114}
\alias{clatworthy.r115}
\alias{clatworthy.r117}
\alias{clatworthy.r118}
\alias{clatworthy.r119}
\alias{clatworthy.r13}
\alias{clatworthy.r133}
\alias{clatworthy.r134}
\alias{clatworthy.r137}
\alias{clatworthy.r139}
\alias{clatworthy.r140}
\alias{clatworthy.r143}
\alias{clatworthy.r144}
\alias{clatworthy.r145}
\alias{clatworthy.r15}
\alias{clatworthy.r150}
\alias{clatworthy.r153}
\alias{clatworthy.r153.}
\alias{clatworthy.r16}
\alias{clatworthy.r164}
\alias{clatworthy.r165}
\alias{clatworthy.r169}
\alias{clatworthy.r17}
\alias{clatworthy.r170}
\alias{clatworthy.r171}
\alias{clatworthy.r172}
\alias{clatworthy.r175}
\alias{clatworthy.r176}
\alias{clatworthy.r177}
\alias{clatworthy.r18}
\alias{clatworthy.r186}
\alias{clatworthy.r19}
\alias{clatworthy.r193}
\alias{clatworthy.r2}
\alias{clatworthy.r20}
\alias{clatworthy.r203}
\alias{clatworthy.r21}
\alias{clatworthy.r22}
\alias{clatworthy.r24}
\alias{clatworthy.r25}
\alias{clatworthy.r26}
\alias{clatworthy.r27}
\alias{clatworthy.r29}
\alias{clatworthy.r3}
\alias{clatworthy.r30}
\alias{clatworthy.r31}
\alias{clatworthy.r32}
\alias{clatworthy.r34}
\alias{clatworthy.r35}
\alias{clatworthy.r36}
\alias{clatworthy.r37}
\alias{clatworthy.r38}
\alias{clatworthy.r39}
\alias{clatworthy.r4}
\alias{clatworthy.r40}
\alias{clatworthy.r41}
\alias{clatworthy.r42}
\alias{clatworthy.r43}
\alias{clatworthy.r44}
\alias{clatworthy.r45}
\alias{clatworthy.r46}
\alias{clatworthy.r47}
\alias{clatworthy.r48}
\alias{clatworthy.r5}
\alias{clatworthy.r50}
\alias{clatworthy.r51}
\alias{clatworthy.r52}
\alias{clatworthy.r54}
\alias{clatworthy.r56}
\alias{clatworthy.r58}
\alias{clatworthy.r59}
\alias{clatworthy.r6}
\alias{clatworthy.r60}
\alias{clatworthy.r61}
\alias{clatworthy.r62}
\alias{clatworthy.r63}
\alias{clatworthy.r65}
\alias{clatworthy.r66}
\alias{clatworthy.r67}
\alias{clatworthy.r68}
\alias{clatworthy.r69}
\alias{clatworthy.r7}
\alias{clatworthy.r70}
\alias{clatworthy.r71}
\alias{clatworthy.r72}
\alias{clatworthy.r73}
\alias{clatworthy.r75}
\alias{clatworthy.r78}
\alias{clatworthy.r79}
\alias{clatworthy.r80}
\alias{clatworthy.r81}
\alias{clatworthy.r82}
\alias{clatworthy.r84}
\alias{clatworthy.r85}
\alias{clatworthy.r86}
\alias{clatworthy.r89}
\alias{clatworthy.r91}
\alias{clatworthy.r93}
\alias{clatworthy.r94}
\alias{clatworthy.r95}
\alias{clatworthy.r96}
\alias{clatworthy.r97}
\alias{clatworthy.r98}
\alias{clatworthy.s1}
\alias{clatworthy.s101}
\alias{clatworthy.s106}
\alias{clatworthy.s11}
\alias{clatworthy.s110}
\alias{clatworthy.s116}
\alias{clatworthy.s18}
\alias{clatworthy.s21}
\alias{clatworthy.s24}
\alias{clatworthy.s26}
\alias{clatworthy.s27}
\alias{clatworthy.s28}
\alias{clatworthy.s29}
\alias{clatworthy.s3}
\alias{clatworthy.s31}
\alias{clatworthy.s32}
\alias{clatworthy.s34}
\alias{clatworthy.s35}
\alias{clatworthy.s37}
\alias{clatworthy.s38}
\alias{clatworthy.s51}
\alias{clatworthy.s52}
\alias{clatworthy.s53}
\alias{clatworthy.s58}
\alias{clatworthy.s59}
\alias{clatworthy.s6}
\alias{clatworthy.s61}
\alias{clatworthy.s63}
\alias{clatworthy.s65}
\alias{clatworthy.s66}
\alias{clatworthy.s67}
\alias{clatworthy.s7}
\alias{clatworthy.s82}
\alias{clatworthy.s85}
\alias{clatworthy.s86}
\alias{clatworthy.s88}
\alias{clatworthy.s9}
\alias{clatworthy.s91}
\alias{clatworthy.s98}
\alias{clatworthy.sr1}
\alias{clatworthy.sr100}
\alias{clatworthy.sr106}
\alias{clatworthy.sr109}
\alias{clatworthy.sr11}
\alias{clatworthy.sr18}
\alias{clatworthy.sr19}
\alias{clatworthy.sr23}
\alias{clatworthy.sr24}
\alias{clatworthy.sr25}
\alias{clatworthy.sr26}
\alias{clatworthy.sr3}
\alias{clatworthy.sr35}
\alias{clatworthy.sr36}
\alias{clatworthy.sr36a}
\alias{clatworthy.sr37}
\alias{clatworthy.sr41}
\alias{clatworthy.sr42}
\alias{clatworthy.sr44}
\alias{clatworthy.sr46}
\alias{clatworthy.sr52}
\alias{clatworthy.sr53}
\alias{clatworthy.sr54}
\alias{clatworthy.sr55}
\alias{clatworthy.sr56}
\alias{clatworthy.sr6}
\alias{clatworthy.sr60}
\alias{clatworthy.sr65}
\alias{clatworthy.sr66}
\alias{clatworthy.sr67}
\alias{clatworthy.sr68}
\alias{clatworthy.sr7}
\alias{clatworthy.sr72}
\alias{clatworthy.sr73}
\alias{clatworthy.sr75}
\alias{clatworthy.sr80}
\alias{clatworthy.sr81}
\alias{clatworthy.sr82}
\alias{clatworthy.sr9}
\alias{clatworthy.sr90}
\alias{clatworthy.sr91}
\alias{clatworthy.sr92}
\alias{clatworthy.sr93}
\alias{clatworthy.sr94}
\alias{clatworthy.sr99}
\alias{pb2.64}
\alias{pb2.williams4t.r101}
\alias{pb2.williams4t.r103}
\alias{pb2.williams4t.r104}
\alias{pb2.williams4t.r106}
\alias{pb2.williams4t.r109}
\alias{pb2.williams4t.r111}
\alias{pb2.williams4t.r112}
\alias{pb2.williams4t.r114}
\alias{pb2.williams4t.r115}
\alias{pb2.williams4t.r117}
\alias{pb2.williams4t.r118}
\alias{pb2.williams4t.r119}
\alias{pb2.williams5t.r133}
\alias{pb2.williams5t.r134}
\alias{pb2.williams5t.r137}
\alias{pb2.williams5t.r139}
\alias{pb2.williams5t.r140}
\alias{pb2.williams5t.r143}
\alias{pb2.williams5t.r144}
\alias{pb2.williams5t.r145}
\alias{pb2.williams5t.r150}
\alias{pb2.williams5t.r153}
\alias{pb2.williams5t.r153~}
\alias{pb2.williams6t.r164}
\alias{pb2.williams6t.r165}
\alias{pb2.williams6t.r169}
\alias{pb2.williams6t.r170}
\alias{pb2.williams6t.r171}
\alias{pb2.williams7t.r172}
\alias{pb2.williams7t.r175}
\alias{pb2.williams7t.r176}
\alias{pb2.williams7t.r177}
\alias{pb2.williams8t.r186}
\alias{pb2.williams9t.r193}
\alias{pb2.williams3t.r42}
\alias{pb2.williams3t.r43}
\alias{pb2.williams3t.r44}
\alias{pb2.williams3t.r45}
\alias{pb2.williams3t.r46}
\alias{pb2.williams3t.r47}
\alias{pb2.williams3t.r48}
\alias{pb2.williams3t.r50}
\alias{pb2.williams3t.r51}
\alias{pb2.williams3t.r52}
\alias{pb2.williams3t.r54}
\alias{pb2.williams3t.r56}
\alias{pb2.williams3t.r58}
\alias{pb2.williams3t.r59}
\alias{pb2.williams3t.r60}
\alias{pb2.williams3t.r61}
\alias{pb2.williams3t.r62}
\alias{pb2.williams3t.r63}
\alias{pb2.williams3t.r65}
\alias{pb2.williams3t.r66}
\alias{pb2.williams3t.r67}
\alias{pb2.williams3t.r68}
\alias{pb2.williams3t.r69}
\alias{pb2.williams3t.r70}
\alias{pb2.williams3t.r71}
\alias{pb2.williams3t.r72}
\alias{pb2.williams3t.r73}
\alias{pb2.williams3t.r75}
\alias{pb2.williams3t.r78}
\alias{pb2.williams3t.r79}
\alias{pb2.williams3t.r80}
\alias{pb2.williams3t.r81}
\alias{pb2.williams3t.r82}
\alias{pb2.williams3t.r84}
\alias{pb2.williams3t.r85}
\alias{pb2.williams3t.r86}
\alias{pb2.williams3t.r89}
\alias{pb2.williams3t.r91}
\alias{pb2.williams3t.r93}
\alias{pb2.williams4t.r94}
\alias{pb2.williams4t.r95}
\alias{pb2.williams4t.r96}
\alias{pb2.williams4t.r97}
\alias{pb2.williams4t.r98}
\alias{pb2.williams4t.s1}
\alias{pb2.williams4t.s11}
\alias{pb2.williams6t.s18}
\alias{pb2.williams6t.s21}
\alias{pb2.williams6t.s24}
\alias{pb2.williams6t.s26}
\alias{pb2.williams6t.s27}
\alias{pb2.williams6t.s28}
\alias{pb2.williams6t.s29}
\alias{pb2.williams4t.s3}
\alias{pb2.williams6t.s31}
\alias{pb2.williams6t.s32}
\alias{pb2.williams6t.s34}
\alias{pb2.williams6t.s35}
\alias{pb2.williams6t.s37}
\alias{pb2.williams6t.s38}
\alias{pb2.williams8t.s51}
\alias{pb2.williams8t.s52}
\alias{pb2.williams8t.s53}
\alias{pb2.williams8t.s58}
\alias{pb2.williams8t.s59}
\alias{pb2.williams4t.s6}
\alias{pb2.williams8t.s61}
\alias{pb2.williams8t.s63}
\alias{pb2.williams8t.s65}
\alias{pb2.williams8t.s66}
\alias{pb2.williams8t.s67}
\alias{pb2.williams4t.s7}
\alias{pb2.williams9t.s82}
\alias{pb2.williams9t.s85}
\alias{pb2.williams9t.s86}
\alias{pb2.williams9t.s88}
\alias{pb2.williams4t.s9}
\alias{pb2.williams9t.s91}
\alias{pb2.williams9t.sr100}
\alias{pb2.williams3t.sr18}
\alias{pb2.williams3t.sr19}
\alias{pb2.williams3t.sr23}
\alias{pb2.williams3t.sr24}
\alias{pb2.williams3t.sr25}
\alias{pb2.williams3t.sr26}
\alias{pb2.williams4t.sr35}
\alias{pb2.williams4t.sr36}
\alias{pb2.williams6t.sr36a}
\alias{pb2.williams4t.sr37}
\alias{pb2.williams4t.sr41}
\alias{pb2.williams4t.sr42}
\alias{pb2.williams4t.sr44}
\alias{pb2.williams4t.sr46}
\alias{pb2.williams5t.sr52}
\alias{pb2.williams5t.sr53}
\alias{pb2.williams5t.sr54}
\alias{pb2.williams5t.sr55}
\alias{pb2.williams5t.sr56}
\alias{pb2.williams5t.sr60}
\alias{pb2.williams6t.sr65}
\alias{pb2.williams6t.sr66}
\alias{pb2.williams6t.sr67}
\alias{pb2.williams6t.sr68}
\alias{pb2.williams6t.sr72}
\alias{pb2.williams6t.sr73}
\alias{pb2.williams6t.sr75}
\alias{pb2.williams7t.sr80}
\alias{pb2.williams7t.sr81}
\alias{pb2.williams7t.sr82}
\alias{pb2.williams8t.sr90}
\alias{pb2.williams8t.sr91}
\alias{pb2.williams8t.sr92}
\alias{pb2.williams8t.sr93}
\alias{pb2.williams8t.sr94}
\alias{pb2.williams9t.sr99}
\title{Selected Cross-Over designs from literature}
\format{A integer matrix specifying the design. Rows represent periods and columns the subjects.}
\source{
Anderson, I. and Preece, D.A. (2002) Locally balanced change-over
designs. Utilitas Mathematica, To appear.

Anderson, I. (2002) Personal communication.

Archdeacon, D.S., Dinitz J.H., Stinson, D.R. and Tillson, T.W. (1980) Some
new row-complete latin squares, Journal of Combinatorial Theory, Series A,
29, 395-398.

Atkinson, G.F. (1966) Designs for sequences of treatments with carry-over
effects. Biometrics, 22, 292--309.

Balaam, L.N. (1968) A two-period design with \eqn{t^2}{t^2} experimental
units. Biometrics, 24, 61--73.

Bate, S. and Jones, B. (2002) The construction of universally optimal
uniform cross-over designs. GlaxoSmithKline Biomedical Data Sciences
Technical Report 2002-06.

Berenblut, I.I. (1964) Change-over designs with complete balance for
residual effects. Biometrics, 23, 578--580.

Blaisdell, E.A. and Raghavarao, D. (1980) Partially balanced change-over
designs based on m-associate class PBIB designs. Journal of the Royal
Statistical Society, B, 42, 334--338.

Clatworthy, W. H. (1973). Tables of two-associate-class partially balanced 
designs. US Government Printing Office.

Davis, A.W. and Hall, W.B. (1969) Cyclic change-over designs. Biometrika,
56, 283--293.

Federer, W.T. and Atkinson, G.F. (1964) Tied-double-change-over designs.
Biometrics, 20, 168--181.

Fletcher, D.J. (1987) A new class of change-over designs for factorial
experiments. Biometrika, 74, 649--654.

Iqbal, I. and Jones, B. (1994) Efficient repeated measurements designs with
equal and unequal period sizes. Journal of Statistical Planning and
Inference, 42, 79-88.

Factorial cross-over designs in clinical trials, Lewis, S.M., Fletcher, D.J.
and Matthews, J.N.S. In Optimal Design and Analysis of Experiments, Editors,
Dodge, Y., Fedorov, V.V. and Wynn, H.P. (1988), 133--140, Elsevier Science
Publishers B.V. (North-Holland).

Cochran, W.G., Autrey, K.M. and Cannon, C.Y. (1941) A double change-over
design for dairy cattle feeding experiments. Journal of Dairy Science, 24,
937--951

Patterson, H.D. and Lucas, H.L. (1962) Change-over designs. North Carolina
Agricultural Experiment Station. Tech. Bull. No. 147.

Residual effects designs for comparing treatments with a control. PhD
dissertation, Temple University, Phildelphia, PA, 1984.

Prescott, P. (1994) Construction of sequentially counterbalanced designs
formed from two or more Latin Squares. Proceedings of the 11th Symposium on
Computational Statistics held in Vienna, Austria. Editors Dutter, R. and
Grossmann, W., Physica-Verlag: Heidelberg, 435-440.

Prescott, P. (1999) Construction of sequentially counterbalanced designs
formed from two latin squares. Utilitas Mathematica, 55, 135--152.

Quenouille, M.H. (1953) The Design and Analysis of Experiments. Griffin,
London.

Russell, K.R. (1991) The construction of good change-over designs when there
are fewer units than treatments. Biometrika, 78, 305-313.

Lucas, H.L. (1956) Switch-back trials for more than two treatments. Journal
of Diary Science, 39, 146--154.

Williams, E.J. (1949) Experimental designs balanced for the estimation of
residual effects of treatments. Australian Journal of Science Res(A), 2,
14900168.
}
\description{
Selected Cross-Over designs from literature. 

You can access all designs via the function \code{getDesign} as in the 
example \code{getDesign("williams4t")}.
}
\details{
These data sets are stored combined by prefix, so alternatively to using the
recommended function \code{getDesign} you coud access for example design
\code{fletcher10} by using the command \code{data(fletcher10)} and
afterwards all 31 design from fletcher1 up to fletcher31 are loaded.

%Another way to get these data sets is to use the function 

The available data sets are:\cr\cr
federerAtkinson3ta, federerAtkinson3tb, federerAtkinson4ta,
federerAtkinson4tb, federerAtkinson5ta, fletcher1, fletcher10, fletcher11,
fletcher12, fletcher13, fletcher14, fletcher15, fletcher16, fletcher17,
fletcher18, fletcher19, fletcher2, fletcher20, fletcher21, fletcher22,
fletcher23, fletcher24, fletcher25, fletcher26, fletcher27, fletcher28,
fletcher29, fletcher3, fletcher30, fletcher31, fletcher4, fletcher5,
fletcher6, fletcher7, fletcher8, fletcher9, iqbalJones1, iqbalJones10,
iqbalJones11, iqbalJones12, iqbalJones13, iqbalJones14, iqbalJones15,
iqbalJones16, iqbalJones17, iqbalJones18, iqbalJones19, iqbalJones2,
iqbalJones20, iqbalJones21, iqbalJones22, iqbalJones23, iqbalJones24,
iqbalJones25, iqbalJones26, iqbalJones27, iqbalJones28, iqbalJones29,
iqbalJones3, iqbalJones30, iqbalJones31, iqbalJones32, iqbalJones33,
iqbalJones34, iqbalJones35, iqbalJones36, iqbalJones37, iqbalJones38,
iqbalJones39, iqbalJones4, iqbalJones40, iqbalJones41, iqbalJones42,
iqbalJones5, iqbalJones6, iqbalJones7, iqbalJones8, iqbalJones9,
lewisFletcherMatthews1, lewisFletcherMatthews10, lewisFletcherMatthews11,
lewisFletcherMatthews12, lewisFletcherMatthews13, lewisFletcherMatthews14,
lewisFletcherMatthews15, lewisFletcherMatthews16, lewisFletcherMatthews17,
lewisFletcherMatthews18, lewisFletcherMatthews19, lewisFletcherMatthews2,
lewisFletcherMatthews20, lewisFletcherMatthews3, lewisFletcherMatthews4,
lewisFletcherMatthews5, lewisFletcherMatthews6, lewisFletcherMatthews7,
lewisFletcherMatthews8, lewisFletcherMatthews9, orthogonalLatinSquare3t,
orthogonalLatinSquare4t, orthogonalLatinSquare5t, orthogonalLatinSquare7t,
pattersonLucasExtraPeriod30, pattersonLucasExtraPeriod31,
pattersonLucasExtraPeriod32, pattersonLucasExtraPeriod33,
pattersonLucasExtraPeriod34, pattersonLucasExtraPeriod35,
pattersonLucasExtraPeriod36, pattersonLucasExtraPeriod37,
pattersonLucasExtraPeriod38, pattersonLucasExtraPeriod39,
pattersonLucasExtraPeriod40, pattersonLucasExtraPeriod41,
pattersonLucasExtraPeriod42, pattersonLucasExtraPeriod43,
pattersonLucasExtraPeriod44, pattersonLucasExtraPeriod45,
pattersonLucasExtraPeriod46, pattersonLucasExtraPeriod47,
pattersonLucasExtraPeriod48, pattersonLucasExtraPeriod49,
pattersonLucasExtraPeriod86, pattersonLucasPBIBD100, pattersonLucasPBIBD101,
pattersonLucasPBIBD102, pattersonLucasPBIBD103, pattersonLucasPBIBD104,
pattersonLucasPBIBD105, pattersonLucasPBIBD106, pattersonLucasPBIBD107,
pattersonLucasPBIBD125, pattersonLucasPBIBD126, pattersonLucasPBIBD127,
pattersonLucasPBIBD128, pattersonLucasPBIBD131, pattersonLucasPBIBD132,
pattersonLucasPBIBD133, pattersonLucasPBIBD134, pattersonLucasPBIBD135,
pattersonLucasPBIBD136, pattersonLucasPBIBD137, pattersonLucasPBIBD138,
pattersonLucasPBIBD139, pattersonLucasPBIBD140, pattersonLucasPBIBD141,
pattersonLucasPBIBD153, pattersonLucasPBIBD154, pattersonLucasPBIBD155,
pattersonLucasPBIBD156, pattersonLucasPBIBD99, pattersonLucasPltT1,
pattersonLucasPltT10, pattersonLucasPltT12, pattersonLucasPltT13,
pattersonLucasPltT15, pattersonLucasPltT16, pattersonLucasPltT17,
pattersonLucasPltT18, pattersonLucasPltT19, pattersonLucasPltT20,
pattersonLucasPltT21, pattersonLucasPltT22, pattersonLucasPltT23,
pattersonLucasPltT3, pattersonLucasPltT4, pattersonLucasPltT5,
pattersonLucasPltT7, pattersonLucasPltT8, pattersonLucasPltT9, pidgeon1,
pidgeon10, pidgeon11, pidgeon12, pidgeon13, pidgeon14, pidgeon15, pidgeon16,
pidgeon17, pidgeon18, pidgeon19, pidgeon2, pidgeon20, pidgeon3, pidgeon4,
pidgeon5, pidgeon6, pidgeon7, pidgeon8, pidgeon9, prescott1, prescott2,
quenouille3t1, quenouille3t2, quenouille4t1, quenouille4t2, quenouille4t3,
russel4t, russel7t, switchback3t, switchback4t, switchback5t, switchback6t,
switchback7t, williams3t, williams4t, williams5t, williams6t, williams7t,
williams8t, williams9t, pb2.64.
}
\examples{

getDesign("williams4t")

data(fletcher)
ls(pattern="fletcher*")
fletcher10
}
%\references{
% source - Details of the original source (a reference or URL). In addition, section \references could give secondary sources and usages.
% references - A section with references to the literature. 
%}
\keyword{datasets}

