% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/startGUIs.R
\name{CrossoverGUI}
\alias{CrossoverGUI}
\title{Graphical User Interface for Crossover Designs}
\usage{
CrossoverGUI()
}
\value{
The function itself returns nothing of interest. But from the GUI designs
and objects can be created or edited that will be available in R under the
specified variable name after saving.
}
\description{
Starts a graphical user interface for accessing and creating crossover designs.
}
\details{
See the vignette of this package for further details, since describing a GUI
interface is better done with some nice pictures.
}
\examples{
\dontrun{
CrossoverGUI()
}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{misc}

