\name{crossoverdesigns}
\docType{data}
\alias{anderson}
\alias{andersonPreece}
\alias{archdeacon}
\alias{atkinson3t}
\alias{atkinson4t}
\alias{atkinson5t}
\alias{balaam3t}
\alias{balaam4t}
\alias{balaam5t}
\alias{balaam6t}
\alias{bateJones5t}
\alias{bateJones8t}
\alias{berenblut3t}
\alias{berenblut4t}
\alias{berenblut5t}
\alias{blaisdellRaghavarao6t}
\alias{blaisdellRaghavarao8t}
\alias{blaisdellRaghavarao9t}
\alias{davisHall6ta}
\alias{davisHall6tb}
\alias{davisHall6tc}
\alias{davisHall7ta}
\alias{davisHall7tb}
\alias{davisHall7tc}
\alias{davisHall8ta}
\alias{davisHall8tb}
\alias{davisHall8tc}
\alias{davisHall9ta}
\alias{davisHall9tb}
\alias{davisHall9tc}
\title{Selected Cross-Over designs from literature}
\description{
	Selected Cross-Over designs from literature.
}
\usage{data(anderson)}
\format{A integer matrix ...:
\describe{
    \item{cross-over design}{All designs are integer matrices. ...}
  }
}
\source{  
  ...
}
\references{
  ...
}
\examples{
data(anderson)
}
\keyword{datasets}

