% Generated by roxygen2 (4.0.1): do not edit by hand
\name{general.carryover}
\alias{general.carryover}
\title{Search for a Cross-Over Design}
\usage{
general.carryover(design, model, t0 = 1, rho = 0.5)
}
\arguments{
\item{design}{Cross-over design.}

\item{model}{Model - one of the following numbers or Strings: 1 = "Standard
additive model", 2 = "Self-adjacency model", 3 = "Proportionality model", 4
= "Placebo model", 5 = "No carry-over into self model", 6 = "Treatment decay
model", 7 = "Full set of interactions", 8 = "Second-order carry-over
effects"}

\item{t0}{The number of placebo treatments in the placebo model.}

\item{rho}{The proportionality parameter for the proportionality model.}
}
\value{
TODO ...
}
\description{
Search for a Cross-Over Design
}
\details{
See the vignette of this package for further details.
}
\examples{
x <- searchCrossOverDesign(s=9, p=5, v=4, model=4, eff.factor=1)
}
\author{
Byron Jones
}
\references{
Jones, B., & Kenward, M. G. (2003). Design and analysis of
cross-over trials (Vol. 98). Chapman & Hall.
}
\keyword{misc}

