\name{nvm-package}
\alias{nvm-package}
\docType{package}
\title{
A graphical mask
}
\description{
This package provides a mask
}
\author{
%Kornelius Rohmeyer 

Maintainer: Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{ package }
\keyword{ graphs }
\keyword{ htest }
%\seealso{
%\code{\link[mutoss:Mutoss]{Mutoss}}
%}
