\docType{data}
\name{exampleSearchResults2t}
\alias{exampleSearchResults2t}
\alias{resultL}
\title{Example search results for two treatments}
\format{A list of 16 integer matrices specifying the design. Rows represent periods and columns the subjects.}
\source{
Found by method searchCrossOverDesign.
}
\description{
A list of 16 integer matrices specifying the design. Rows represent periods and columns the subjects.
}
\details{
See vignette.
}
\keyword{datasets}

