\name{crossoverGUI}
\alias{crossoverGUI}

\title{
	Graphical User Interface for the Creation of SRMTP Graphs
}
\description{
	Starts a graphical user interface for the creation of SRMTP graphs.
}
\usage{
	crossoverGUI()
}

\details{
	See the vignette of this package for further details, since describing a GUI interface is better done with a lot of nice pictures.
}
\value{
	The function itself returns NULL. But from the GUI designs and objects can be created or edited that will be available in R under the specified variable name after saving.
}
\author{
	Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

\keyword{ misc }
\keyword{ graphs }

\examples{

\dontrun{
crossoverGUI()
}

}