\name{buildSummaryTable}
\alias{buildSummaryTable}
\title{Build Summary Table For All Examples From Literature}
\usage{
  buildSummaryTable()
}
\value{
  TODO
}
\description{
  Build Summary Table For All Examples From Literature
}
\details{
  See also the documentation for the data files.
}
\examples{
buildSummaryTable()
}
\author{
  Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\references{
  See the documentation for the data files.
}
\keyword{misc}

