% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildSummary.R
\name{buildSummaryTable}
\alias{buildSummaryTable}
\title{Build Summary Table For All Examples From Literature}
\usage{
buildSummaryTable(extended = FALSE)
}
\arguments{
\item{extended}{If \code{TRUE} the summary table will have further
columns with extended information as how balanced the design is and
whether all treatment effect differences are estimable under all
models.}
}
\value{
TODO
}
\description{
Build Summary Table For All Examples From Literature
}
\details{
See also the documentation for the data files.
}
\examples{
buildSummaryTable()
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\references{
See the documentation for the data files.
}
\keyword{misc}

