\name{crossover-package}
\alias{crossover-package}
\docType{package}
\title{
A graphical mask
}
\description{
This package provides a mask
}
\author{
%Kornelius Rohmeyer 

Maintainer: Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\references{John, J. A., Russell, K. G., & Whitaker, D. (2004). CrossOver: an algorithm for the construction of efficient cross‐over designs. Statistics in medicine, 23(17), 2645-2658.}
\keyword{ package }
\keyword{ graphs }
\keyword{ htest }
%\seealso{
%\code{\link[mutoss:Mutoss]{Mutoss}}
%}
