\docType{package}
\name{Crossover-package}
\alias{Crossover-package}
\title{This package provides more than two hundred cross-over design from
literature, a search algorithm to find efficient cross-over designs for
various models and a graphical user interface to find/generate appropriate
designs.}
\description{
  This package provides more than two hundred cross-over
  design from literature, a search algorithm to find
  efficient cross-over designs for various models and a
  graphical user interface to find/generate appropriate
  designs.
}
\author{
  %Kornelius Rohmeyer, Byron Jones

  Maintainer: Kornelius Rohmeyer
  \email{rohmeyer@small-projects.de}
}
\references{
  Jones, B., & Kenward, M. G. (2003). Design and analysis
  of cross-over trials (Vol. 98). Chapman & Hall.

  John, J. A., Russell, K. G., & Whitaker, D. (2004).
  CrossOver: an algorithm for the construction of efficient
  cross-over designs. Statistics in medicine, 23(17),
  2645-2658.
}
\keyword{htest}
\keyword{package}

