\docType{data}
\name{crossoverdesigns}
\alias{anderson}
\alias{andersonPreece}
\alias{archdeacon}
\alias{atkinson3t}
\alias{atkinson4t}
\alias{atkinson5t}
\alias{balaam3t}
\alias{balaam4t}
\alias{balaam5t}
\alias{balaam6t}
\alias{bateJones5t}
\alias{bateJones8t}
\alias{berenblut3t}
\alias{berenblut4t}
\alias{berenblut5t}
\alias{blaisdellRaghavarao6t}
\alias{blaisdellRaghavarao8t}
\alias{blaisdellRaghavarao9t}
\alias{crossoverdesigns}
\alias{davisHall6ta}
\alias{davisHall6tb}
\alias{davisHall6tc}
\alias{davisHall7ta}
\alias{davisHall7tb}
\alias{davisHall7tc}
\alias{davisHall8ta}
\alias{davisHall8tb}
\alias{davisHall8tc}
\alias{davisHall9ta}
\alias{davisHall9tb}
\alias{davisHall9tc}
\alias{federerAtkinson3ta}
\alias{federerAtkinson3tb}
\alias{federerAtkinson4ta}
\alias{federerAtkinson4tb}
\alias{federerAtkinson5ta}
\alias{fletcher1}
\alias{fletcher10}
\alias{fletcher11}
\alias{fletcher12}
\alias{fletcher13}
\alias{fletcher14}
\alias{fletcher15}
\alias{fletcher16}
\alias{fletcher17}
\alias{fletcher18}
\alias{fletcher19}
\alias{fletcher2}
\alias{fletcher20}
\alias{fletcher21}
\alias{fletcher22}
\alias{fletcher23}
\alias{fletcher24}
\alias{fletcher25}
\alias{fletcher26}
\alias{fletcher27}
\alias{fletcher28}
\alias{fletcher29}
\alias{fletcher3}
\alias{fletcher30}
\alias{fletcher31}
\alias{fletcher4}
\alias{fletcher5}
\alias{fletcher6}
\alias{fletcher7}
\alias{fletcher8}
\alias{fletcher9}
\alias{iqbalJones1}
\alias{iqbalJones10}
\alias{iqbalJones11}
\alias{iqbalJones12}
\alias{iqbalJones13}
\alias{iqbalJones14}
\alias{iqbalJones15}
\alias{iqbalJones16}
\alias{iqbalJones17}
\alias{iqbalJones18}
\alias{iqbalJones19}
\alias{iqbalJones2}
\alias{iqbalJones20}
\alias{iqbalJones21}
\alias{iqbalJones22}
\alias{iqbalJones23}
\alias{iqbalJones24}
\alias{iqbalJones25}
\alias{iqbalJones26}
\alias{iqbalJones27}
\alias{iqbalJones28}
\alias{iqbalJones29}
\alias{iqbalJones3}
\alias{iqbalJones30}
\alias{iqbalJones31}
\alias{iqbalJones32}
\alias{iqbalJones33}
\alias{iqbalJones34}
\alias{iqbalJones35}
\alias{iqbalJones36}
\alias{iqbalJones37}
\alias{iqbalJones38}
\alias{iqbalJones39}
\alias{iqbalJones4}
\alias{iqbalJones40}
\alias{iqbalJones41}
\alias{iqbalJones42}
\alias{iqbalJones5}
\alias{iqbalJones6}
\alias{iqbalJones7}
\alias{iqbalJones8}
\alias{iqbalJones9}
\alias{lewisFletcherMatthews1}
\alias{lewisFletcherMatthews10}
\alias{lewisFletcherMatthews11}
\alias{lewisFletcherMatthews12}
\alias{lewisFletcherMatthews13}
\alias{lewisFletcherMatthews14}
\alias{lewisFletcherMatthews15}
\alias{lewisFletcherMatthews16}
\alias{lewisFletcherMatthews17}
\alias{lewisFletcherMatthews18}
\alias{lewisFletcherMatthews19}
\alias{lewisFletcherMatthews2}
\alias{lewisFletcherMatthews20}
\alias{lewisFletcherMatthews3}
\alias{lewisFletcherMatthews4}
\alias{lewisFletcherMatthews5}
\alias{lewisFletcherMatthews6}
\alias{lewisFletcherMatthews7}
\alias{lewisFletcherMatthews8}
\alias{lewisFletcherMatthews9}
\alias{orthogonalLatinSquare3t}
\alias{orthogonalLatinSquare4t}
\alias{orthogonalLatinSquare5t}
\alias{orthogonalLatinSquare7t}
\alias{pattersonLucasExtraPeriod30}
\alias{pattersonLucasExtraPeriod31}
\alias{pattersonLucasExtraPeriod32}
\alias{pattersonLucasExtraPeriod33}
\alias{pattersonLucasExtraPeriod34}
\alias{pattersonLucasExtraPeriod35}
\alias{pattersonLucasExtraPeriod36}
\alias{pattersonLucasExtraPeriod37}
\alias{pattersonLucasExtraPeriod38}
\alias{pattersonLucasExtraPeriod39}
\alias{pattersonLucasExtraPeriod40}
\alias{pattersonLucasExtraPeriod41}
\alias{pattersonLucasExtraPeriod42}
\alias{pattersonLucasExtraPeriod43}
\alias{pattersonLucasExtraPeriod44}
\alias{pattersonLucasExtraPeriod45}
\alias{pattersonLucasExtraPeriod46}
\alias{pattersonLucasExtraPeriod47}
\alias{pattersonLucasExtraPeriod48}
\alias{pattersonLucasExtraPeriod49}
\alias{pattersonLucasExtraPeriod86}
\alias{pattersonLucasPBIBD100}
\alias{pattersonLucasPBIBD101}
\alias{pattersonLucasPBIBD102}
\alias{pattersonLucasPBIBD103}
\alias{pattersonLucasPBIBD104}
\alias{pattersonLucasPBIBD105}
\alias{pattersonLucasPBIBD106}
\alias{pattersonLucasPBIBD107}
\alias{pattersonLucasPBIBD125}
\alias{pattersonLucasPBIBD126}
\alias{pattersonLucasPBIBD127}
\alias{pattersonLucasPBIBD128}
\alias{pattersonLucasPBIBD131}
\alias{pattersonLucasPBIBD132}
\alias{pattersonLucasPBIBD133}
\alias{pattersonLucasPBIBD134}
\alias{pattersonLucasPBIBD135}
\alias{pattersonLucasPBIBD136}
\alias{pattersonLucasPBIBD137}
\alias{pattersonLucasPBIBD138}
\alias{pattersonLucasPBIBD139}
\alias{pattersonLucasPBIBD140}
\alias{pattersonLucasPBIBD141}
\alias{pattersonLucasPBIBD153}
\alias{pattersonLucasPBIBD154}
\alias{pattersonLucasPBIBD155}
\alias{pattersonLucasPBIBD156}
\alias{pattersonLucasPBIBD99}
\alias{pattersonLucasPltT1}
\alias{pattersonLucasPltT10}
\alias{pattersonLucasPltT12}
\alias{pattersonLucasPltT13}
\alias{pattersonLucasPltT15}
\alias{pattersonLucasPltT16}
\alias{pattersonLucasPltT17}
\alias{pattersonLucasPltT18}
\alias{pattersonLucasPltT19}
\alias{pattersonLucasPltT20}
\alias{pattersonLucasPltT21}
\alias{pattersonLucasPltT22}
\alias{pattersonLucasPltT23}
\alias{pattersonLucasPltT3}
\alias{pattersonLucasPltT4}
\alias{pattersonLucasPltT5}
\alias{pattersonLucasPltT7}
\alias{pattersonLucasPltT8}
\alias{pattersonLucasPltT9}
\alias{pidgeon1}
\alias{pidgeon10}
\alias{pidgeon11}
\alias{pidgeon12}
\alias{pidgeon13}
\alias{pidgeon14}
\alias{pidgeon15}
\alias{pidgeon16}
\alias{pidgeon17}
\alias{pidgeon18}
\alias{pidgeon19}
\alias{pidgeon2}
\alias{pidgeon20}
\alias{pidgeon3}
\alias{pidgeon4}
\alias{pidgeon5}
\alias{pidgeon6}
\alias{pidgeon7}
\alias{pidgeon8}
\alias{pidgeon9}
\alias{prescott1}
\alias{prescott2}
\alias{quenouille3t1}
\alias{quenouille3t2}
\alias{quenouille4t1}
\alias{quenouille4t2}
\alias{quenouille4t3}
\alias{russel4t}
\alias{russel7t}
\alias{switchback3t}
\alias{switchback4t}
\alias{switchback5t}
\alias{switchback6t}
\alias{switchback7t}
\alias{williams3t}
\alias{williams4t}
\alias{williams5t}
\alias{williams6t}
\alias{williams7t}
\alias{williams8t}
\alias{williams9t}
\title{Selected Cross-Over designs from literature}
\format{A integer matrix ...: \describe{ \item{cross-over design}{All
designs are integer matrices. ...} }}
\source{
  Anderson, I. and Preece, D.A. (2002) Locally balanced
  change-over designs. Utilitas Mathematica, To appear.

  Anderson, I. (2002) Personal communication.

  Archdeacon, D.S., Dinitz J.H., Stinson, D.R. and Tillson,
  T.W. (1980) Some new row-complete latin squares, Journal
  of Combinatorial Theory, Series A, 29, 395-398.

  Atkinson, G.F. (1966) Designs for sequences of treatments
  with carry-over effects. Biometrics, 22, 292--309.

  Balaam, L.N. (1968) A two-period design with
  \eqn{t^2}{t^2} experimental units. Biometrics, 24,
  61--73.

  Bate, S. and Jones, B. (2002) The construction of
  universally optimal uniform cross-over designs.
  GlaxoSmithKline Biomedical Data Sciences Technical Report
  2002-06.

  Berenblut, I.I. (1964) Change-over designs with complete
  balance for residual effects. Biometrics, 23, 578--580.

  Blaisdell, E.A. and Raghavarao, D. (1980) Partially
  balanced change-over designs based on m-associate class
  PBIB designs. Journal of the Royal Statistical Society,
  B, 42, 334--338.

  Davis, A.W. and Hall, W.B. (1969) Cyclic change-over
  designs. Biometrika, 56, 283--293.

  Federer, W.T. and Atkinson, G.F. (1964)
  Tied-double-change-over designs. Biometrics, 20,
  168--181.

  Fletcher, D.J. (1987) A new class of change-over designs
  for factorial experiments. Biometrika, 74, 649--654.

  Iqbal, I. and Jones, B. (1994) Efficient repeated
  measurements designs with equal and unequal period sizes.
  Journal of Statistical Planning and Inference, 42, 79-88.

  Factorial cross-over designs in clinical trials, Lewis,
  S.M., Fletcher, D.J. and Matthews, J.N.S. In Optimal
  Design and Analysis of Experiments, Editors, Dodge, Y.,
  Fedorov, V.V. and Wynn, H.P. (1988), 133--140, Elsevier
  Science Publishers B.V. (North-Holland).

  Cochran, W.G., Autrey, K.M. and Cannon, C.Y. (1941) A
  double change-over design for dairy cattle feeding
  experiments. Journal of Dairy Science, 24, 937--951

  Patterson, H.D. and Lucas, H.L. (1962) Change-over
  designs. North Carolina Agricultural Experiment Station.
  Tech. Bull. No. 147.

  Residual effects designs for comparing treatments with a
  control. PhD dissertation, Temple University,
  Phildelphia, PA, 1984.

  Prescott, P. (1994) Construction of sequentially
  counterbalanced designs formed from two or more Latin
  Squares. Proceedings of the 11th Symposium on
  Computational Statistics held in Vienna, Austria. Editors
  Dutter, R. and Grossmann, W., Physica-Verlag: Heidelberg,
  435-440.

  Prescott, P. (1999) Construction of sequentially
  counterbalanced designs formed from two latin squares.
  Utilitas Mathematica, 55, 135--152.

  Quenouille, M.H. (1953) The Design and Analysis of
  Experiments. Griffin, London.

  Russell, K.R. (1991) The construction of good change-over
  designs when there are fewer units than treatments.
  Biometrika, 78, 305-313.

  Lucas, H.L. (1956) Switch-back trials for more than two
  treatments. Journal of Diary Science, 39, 146--154.

  Williams, E.J. (1949) Experimental designs balanced for
  the estimation of residual effects of treatments.
  Australian Journal of Science Res(A), 2, 14900168.
}
\description{
  Selected Cross-Over designs from literature. The
  available data sets are: federerAtkinson3ta,
  federerAtkinson3tb, federerAtkinson4ta,
  federerAtkinson4tb, federerAtkinson5ta, fletcher1,
  fletcher10, fletcher11, fletcher12, fletcher13,
  fletcher14, fletcher15, fletcher16, fletcher17,
  fletcher18, fletcher19, fletcher2, fletcher20,
  fletcher21, fletcher22, fletcher23, fletcher24,
  fletcher25, fletcher26, fletcher27, fletcher28,
  fletcher29, fletcher3, fletcher30, fletcher31, fletcher4,
  fletcher5, fletcher6, fletcher7, fletcher8, fletcher9,
  iqbalJones1, iqbalJones10, iqbalJones11, iqbalJones12,
  iqbalJones13, iqbalJones14, iqbalJones15, iqbalJones16,
  iqbalJones17, iqbalJones18, iqbalJones19, iqbalJones2,
  iqbalJones20, iqbalJones21, iqbalJones22, iqbalJones23,
  iqbalJones24, iqbalJones25, iqbalJones26, iqbalJones27,
  iqbalJones28, iqbalJones29, iqbalJones3, iqbalJones30,
  iqbalJones31, iqbalJones32, iqbalJones33, iqbalJones34,
  iqbalJones35, iqbalJones36, iqbalJones37, iqbalJones38,
  iqbalJones39, iqbalJones4, iqbalJones40, iqbalJones41,
  iqbalJones42, iqbalJones5, iqbalJones6, iqbalJones7,
  iqbalJones8, iqbalJones9, lewisFletcherMatthews1,
  lewisFletcherMatthews10, lewisFletcherMatthews11,
  lewisFletcherMatthews12, lewisFletcherMatthews13,
  lewisFletcherMatthews14, lewisFletcherMatthews15,
  lewisFletcherMatthews16, lewisFletcherMatthews17,
  lewisFletcherMatthews18, lewisFletcherMatthews19,
  lewisFletcherMatthews2, lewisFletcherMatthews20,
  lewisFletcherMatthews3, lewisFletcherMatthews4,
  lewisFletcherMatthews5, lewisFletcherMatthews6,
  lewisFletcherMatthews7, lewisFletcherMatthews8,
  lewisFletcherMatthews9, orthogonalLatinSquare3t,
  orthogonalLatinSquare4t, orthogonalLatinSquare5t,
  orthogonalLatinSquare7t, pattersonLucasExtraPeriod30,
  pattersonLucasExtraPeriod31, pattersonLucasExtraPeriod32,
  pattersonLucasExtraPeriod33, pattersonLucasExtraPeriod34,
  pattersonLucasExtraPeriod35, pattersonLucasExtraPeriod36,
  pattersonLucasExtraPeriod37, pattersonLucasExtraPeriod38,
  pattersonLucasExtraPeriod39, pattersonLucasExtraPeriod40,
  pattersonLucasExtraPeriod41, pattersonLucasExtraPeriod42,
  pattersonLucasExtraPeriod43, pattersonLucasExtraPeriod44,
  pattersonLucasExtraPeriod45, pattersonLucasExtraPeriod46,
  pattersonLucasExtraPeriod47, pattersonLucasExtraPeriod48,
  pattersonLucasExtraPeriod49, pattersonLucasExtraPeriod86,
  pattersonLucasPBIBD100, pattersonLucasPBIBD101,
  pattersonLucasPBIBD102, pattersonLucasPBIBD103,
  pattersonLucasPBIBD104, pattersonLucasPBIBD105,
  pattersonLucasPBIBD106, pattersonLucasPBIBD107,
  pattersonLucasPBIBD125, pattersonLucasPBIBD126,
  pattersonLucasPBIBD127, pattersonLucasPBIBD128,
  pattersonLucasPBIBD131, pattersonLucasPBIBD132,
  pattersonLucasPBIBD133, pattersonLucasPBIBD134,
  pattersonLucasPBIBD135, pattersonLucasPBIBD136,
  pattersonLucasPBIBD137, pattersonLucasPBIBD138,
  pattersonLucasPBIBD139, pattersonLucasPBIBD140,
  pattersonLucasPBIBD141, pattersonLucasPBIBD153,
  pattersonLucasPBIBD154, pattersonLucasPBIBD155,
  pattersonLucasPBIBD156, pattersonLucasPBIBD99,
  pattersonLucasPltT1, pattersonLucasPltT10,
  pattersonLucasPltT12, pattersonLucasPltT13,
  pattersonLucasPltT15, pattersonLucasPltT16,
  pattersonLucasPltT17, pattersonLucasPltT18,
  pattersonLucasPltT19, pattersonLucasPltT20,
  pattersonLucasPltT21, pattersonLucasPltT22,
  pattersonLucasPltT23, pattersonLucasPltT3,
  pattersonLucasPltT4, pattersonLucasPltT5,
  pattersonLucasPltT7, pattersonLucasPltT8,
  pattersonLucasPltT9, pidgeon1, pidgeon10, pidgeon11,
  pidgeon12, pidgeon13, pidgeon14, pidgeon15, pidgeon16,
  pidgeon17, pidgeon18, pidgeon19, pidgeon2, pidgeon20,
  pidgeon3, pidgeon4, pidgeon5, pidgeon6, pidgeon7,
  pidgeon8, pidgeon9, prescott1, prescott2, quenouille3t1,
  quenouille3t2, quenouille4t1, quenouille4t2,
  quenouille4t3, russel4t, russel7t, switchback3t,
  switchback4t, switchback5t, switchback6t, switchback7t,
  williams3t, williams4t, williams5t, williams6t,
  williams7t, williams8t, williams9t.
}
\examples{
data(anderson)
}
\references{
  ...
}
\keyword{datasets}

