\name{CrossoverGUI}
\alias{CrossoverGUI}
\title{Graphical User Interface for the Creation of SRMTP Graphs}
\usage{
  CrossoverGUI()
}
\value{
  The function itself returns NULL. But from the GUI
  designs and objects can be created or edited that will be
  available in R under the specified variable name after
  saving.
}
\description{
  Starts a graphical user interface for the creation of
  SRMTP graphs.
}
\details{
  See the vignette of this package for further details,
  since describing a GUI interface is better done with a
  lot of nice pictures.
}
\examples{
\dontrun{
CrossoverGUI()
}
}
\author{
  Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{misc}

