% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{CrossoverSearchResult-class}
\alias{CrossoverSearchResult}
\alias{CrossoverSearchResult-class}
\alias{print,CrossoverSearchResult-method}
\alias{show,CrossoverSearchResult-method}
\title{Class CrossoverSearchResult}
\description{
A S4 class for the search result for Crossover designs:
CrossoverSearchResult
}
\section{Slots}{
 \describe{ \item{list("design")}{An object of class
\code{CrossoverDesign} describing the best design that was
found.}
\item{list("startDesigns")}{A list of start
designs to search from.}
\item{list("model")}{A numeric specifying the model the design was searched
for or -1 if unknown.}
\item{list("eff")}{List, Progress of the
algorithm. TODO: Explain further.}
\item{list("search")}{List, TODO}
\item{list("time")}{Named numeric with the time in seconds the
algorithm was searching.}
\item{list("misc")}{List - in the moment not used.}}
}
\examples{
# n=c(100,10) is very small, but it's just an example and should not take much time
x <- searchCrossOverDesign(s=9, p=5, v=4, model=4, n=c(100,10))
print(x)
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

