% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDesign}
\alias{getDesign}
\alias{getDesign,CrossoverSearchResult-method}
\alias{getDesign,matrix-method}
\title{Extract Design from a CrossoverSearchResult}
\usage{
\S4method{getDesign}{CrossoverSearchResult}(object, ...)
}
\arguments{
\item{object}{A searchCrossOverDesign object from which the design should be extracted.}

\item{...}{Possible parameters for subclasses (not yet used).}
}
\value{
Returns a numeric matrix representing the crossover design.
}
\description{
Extract Design from a CrossoverSearchResult
}
\examples{
#'
x <- searchCrossOverDesign(s=9, p=5, v=4, model=4)
getDesign(x)
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

