% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crossoverS4.R
\docType{methods}
\name{getDesign}
\alias{getDesign}
\alias{getDesign,CrossoverSearchResult-method}
\alias{getDesign,character-method}
\alias{getDesign,matrix-method}
\title{Extract Design from a CrossoverSearchResult}
\usage{
\S4method{getDesign}{CrossoverSearchResult}(object, ...)
}
\arguments{
\item{object}{A searchCrossOverDesign object from which the design should be extracted.}

\item{...}{Possible parameters for subclasses (not yet used).}
}
\value{
Returns a numeric matrix representing the crossover design.
Rows represent periods, columns represent sequences.
}
\description{
Extract Design from a CrossoverSearchResult
}
\examples{
# n=c(100,10) is very small, but it's just an example and should not take much time
x <- searchCrossOverDesign(s=9, p=5, v=4, model=4, n=c(100,10))
getDesign(x)

getDesign("williams4t")
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

